+++ spring-oxygen-core/src/main/java/com/isxcode/oxygen/core/secret/JwtUtils.java
@@ -3,15 +3,14 @@ package com.isxcode.oxygen.core.secret;
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.isxcode.oxygen.core.exception.OxygenException;
+import io.jsonwebtoken.JwtBuilder;
+import io.jsonwebtoken.JwtParserBuilder;
 import io.jsonwebtoken.Jwts;
 import io.jsonwebtoken.SignatureAlgorithm;
 import io.jsonwebtoken.security.Keys;
 
 import java.security.Key;
-import java.util.Date;
-import java.util.HashMap;
-import java.util.Map;
-import java.util.UUID;
+import java.util.*;
 
 /**
  * JWT utils
@@ -36,7 +35,7 @@ public class JwtUtils {
      * @return jwt String
      * @since 2019-12-12
      */
-    public static String encrypt(String aesKey, Object obj) throws OxygenException {
+    public static String encrypt(String aesKey, Object obj, String jwtKey, Integer minutes) throws OxygenException {
 
         Map<String, Object> claims = new HashMap<>(1);
 
@@ -48,12 +47,26 @@ public class JwtUtils {
             }
             claims.put(SecretConstants.CLAIM_KEY, claimsStr);
 
-            return Jwts.builder()
-                .signWith(key)
+            JwtBuilder jwtBuilder = Jwts.builder();
+            if (jwtKey == null) {
+                jwtBuilder = jwtBuilder.signWith(key);
+            } else {
+                jwtBuilder = jwtBuilder.signWith(Keys.hmacShaKeyFor(Arrays.copyOf(jwtKey.getBytes(), 1 << 5)));
+            }
+
+             jwtBuilder
                 .setClaims(claims)
                 .setIssuedAt(new Date())
-                .setId(String.valueOf(UUID.randomUUID()))
-                .compact();
+                .setId(String.valueOf(UUID.randomUUID()));
+
+            if (minutes > 0) {
+                Calendar calendar = Calendar.getInstance();
+                calendar.setTime(new Date());
+                calendar.add(Calendar.MINUTE, minutes);
+                jwtBuilder = jwtBuilder.setExpiration(calendar.getTime());
+            }
+
+            return jwtBuilder.compact();
 
         } catch (JsonProcessingException e) {
 
@@ -70,7 +83,31 @@ public class JwtUtils {
      */
     public static String encrypt(Object obj) throws OxygenException {
 
-        return encrypt(null, obj);
+        return encrypt(null, obj, null, 0);
+    }
+
+    /**
+     * jwt encrypt
+     *
+     * @param obj object
+     * @return jwt String
+     * @since 2019-12-12
+     */
+    public static String encrypt(Object obj, String jwtKey) throws OxygenException {
+
+        return encrypt(null, obj, jwtKey, 0);
+    }
+
+    /**
+     * jwt encrypt
+     *
+     * @param obj object
+     * @return jwt String
+     * @since 2019-12-12
+     */
+    public static String encrypt(Object obj, String jwtKey,Integer minutes) throws OxygenException {
+
+        return encrypt(null, obj, jwtKey, minutes);
     }
 
     /**
@@ -83,11 +120,17 @@ public class JwtUtils {
      * @return A
      * @since 2019-12-12
      */
-    public static <A> A decrypt(String aesKey, String jwtString, Class<A> targetClass) {
+    public static <A> A decrypt(String aesKey, String jwtString, String jwtKey, Class<A> targetClass) {
 
-        String claimStr = Jwts.parserBuilder()
-            .setSigningKey(key)
-            .build()
+        JwtParserBuilder jwtParserBuilder = Jwts.parserBuilder();
+
+        if (jwtKey == null) {
+            jwtParserBuilder = jwtParserBuilder.setSigningKey(key);
+        }else{
+            jwtParserBuilder = jwtParserBuilder.setSigningKey(Keys.hmacShaKeyFor(Arrays.copyOf(jwtKey.getBytes(), 1 << 5)));
+        }
+
+        String claimStr = jwtParserBuilder.build()
             .parseClaimsJws(jwtString)
             .getBody()
             .get(SecretConstants.CLAIM_KEY, String.class);
@@ -116,6 +159,20 @@ public class JwtUtils {
      */
     public static <A> A decrypt(String jwtString, Class<A> claimClass) {
 
-        return decrypt(null, jwtString, claimClass);
+        return decrypt(null, jwtString, null, claimClass);
+    }
+
+    /**
+     * jwt decript
+     *
+     * @param jwtString  jwt
+     * @param <A>        A
+     * @param claimClass claimClass
+     * @return A
+     * @since 2019-12-12
+     */
+    public static <A> A decrypt(String jwtString, String jwtKey, Class<A> claimClass) {
+
+        return decrypt(null, jwtString, jwtKey, claimClass);
     }
 }